remove all malloc+free pairs
For unescapes, this reduces data copies and in the likely case
where the majority of characters do not need to be unescaped,
memory efficiently is not greatly impacted.

For escapes, this seems to help a small bit on escaping where
the output string is longer, too.  rb_str_resize() will release
memory back to the malloc() pool if the disparity between the
buffer size and string length is large enough (1024 bytes in
Ruby 1.9), so memory efficiency for escapes is not overly
impacted.

The removal of direct malloc() calls (vs xmalloc()) also
protects us (as best as possible) from out-of-memory errors,
but will also cause the GC to be invoked more frequently.